move custom protocol files to user dir
