+++ src/coro_rpc/examples/helloworld/extend_to_support_other_rpc_server/main.cpp
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include <coro_rpc/coro_rpc/protocol/rest_rpc_protocol.hpp>
 #include <coro_rpc/coro_rpc_server.hpp>
 #include <thread>
 
+#include "rest_rpc_protocol.hpp"
 #include "rpc_service.h"
 using namespace coro_rpc;
 using namespace async_simple;
